ConferenceChasor/_reference/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.39.0
openpyxl>=3.1.0
python-docx>=1.1.0
